Mask AWS credentials (#39)
* Mask AWS credentials
__NODOCS__

* move delete to stop (#40)

remove unused names

support external tf binary

+++ cloudify_tf/tasks.py
@@ -128,15 +128,8 @@ def install(ctx, **_):
                         'If you do not have sufficient permissions, that '
                         'installation will fail.'.format(
                             loc=executable_path))
-        installation_zip = os.path.join(installation_dir, 'tf.zip')
-        ctx.logger.info(
-            'Downloading Terraform from {source} into {zip}.'.format(
-                source=installation_source,
-                zip=installation_zip))
-        utils.download_file(installation_zip, installation_source)
-        executable_dir = os.path.dirname(executable_path)
-        utils._unzip_and_set_permissions(installation_zip, executable_dir)
-        os.remove(installation_zip)
+        utils.install_binary(
+            installation_dir, executable_path, installation_source)
 
     # store the values in the runtime for safe keeping -> validation
     ctx.instance.runtime_properties['executable_path'] = executable_path
@@ -151,49 +144,58 @@ def uninstall(ctx, **_):
     exc_path = terraform_config.get('executable_path', '')
     system_exc = resource_config.get('use_existing_resource')
 
-    if os.path.isfile(exc_path) and not system_exc:
-        ctx.logger.info('Removing executable: {path}'.format(path=exc_path))
-        os.remove(exc_path)
-    elif os.path.isfile(exc_path):
-        ctx.logger.warn('Unable to remove file {loc} because it is a system '
-                        'resource.'.format(loc=exc_path))
+    if os.path.isfile(exc_path):
+        if system_exc:
+            ctx.logger.info(
+                'Not removing Terraform installation at {loc} as'
+                'it was provided externally'.format(loc=exc_path))
+        else:
+            ctx.logger.info('Removing executable: {path}'.format(
+                path=exc_path))
+            os.remove(exc_path)
 
     for property_name, property_desc in [
             ('plugins_dir', 'plugins directory'),
             ('storage_path', 'storage_directory')]:
-        dir_to_delete = terraform_config.get(property_name, '')
-        utils.remove_dir(dir_to_delete, property_desc)
+        dir_to_delete = terraform_config.get(property_name, None)
+        if dir_to_delete:
+            utils.remove_dir(dir_to_delete, property_desc)
 
 
 @operation
-@skip_if_existing
 def set_directory_config(ctx, **_):
     exc_path = utils.get_executable_path(target=True)
     plugins_dir = utils.get_plugins_dir(target=True)
     storage_path = utils.get_storage_path(target=True)
+    deployment_terraform_dir = os.path.join(storage_path,
+                                            '.terraform')
     resource_node_instance_dir = utils.get_node_instance_dir(source=True)
     if not os.path.exists(resource_node_instance_dir):
         mkdir_p(resource_node_instance_dir)
     resource_terraform_dir = os.path.join(resource_node_instance_dir,
                                           '.terraform')
-    deployment_terraform_dir = os.path.join(storage_path,
-                                            '.terraform')
-
-    # We don't want to put all the plugins for all the node instances in a
-    # deployment multiple times on the system. So here,
-    # we already stored it once on the file system, and now we create
-    # symlinks so other deployments can use it.
-    # TODO: Possibly put this in "apply" and remove the relationship in
-    # the future.
-
-    ctx.logger.info('Creating link {src} {dst}'.format(
-        src=deployment_terraform_dir, dst=resource_terraform_dir))
-    os.symlink(deployment_terraform_dir, resource_terraform_dir)
-
     resource_plugins_dir = plugins_dir.replace(
         ctx.target.instance.id, ctx.source.instance.id)
     resource_storage_dir = storage_path.replace(
         ctx.target.instance.id, ctx.source.instance.id)
+
+    if utils.is_using_existing(target=True):
+        # We are going to use a TF binary at another location.
+        # However, we still need to make sure that this directory exists.
+        # Otherwise TF will complain. It does not create it.
+        # In our other scenario, a symlink is created.
+        mkdir_p(resource_terraform_dir)
+    else:
+        # We don't want to put all the plugins for all the node instances in a
+        # deployment multiple times on the system. So here,
+        # we already stored it once on the file system, and now we create
+        # symlinks so other deployments can use it.
+        # TODO: Possibly put this in "apply" and remove the relationship in
+        # the future.
+
+        ctx.logger.info('Creating link {src} {dst}'.format(
+            src=deployment_terraform_dir, dst=resource_terraform_dir))
+        os.symlink(deployment_terraform_dir, resource_terraform_dir)
 
     ctx.logger.info("setting executable_path to {path}".format(
         path=exc_path))
